pwtgbot/requirements.txt
loguru
pwsearch
python-telegram-bot